Rename filename using snake_case
