+++ feature_selection/__init__.py
@@ -1,2 +1,2 @@
 __version__ = '0.1.0'
-from feature_selection.RecursiveFeatureElimination import recursive_feature_elimination
+from feature_selection.recursive_feature_elimination import recursive_feature_elimination
